waste-detection-app/backend/requirements.txt
Flask
torch
torchvision
numpy
Pillow
flask-cors
pandas
Ultralytics
tqdm
seaborn